construction-ai-service/requirements.txt
fastapi
uvicorn
firebase-admin
ezdxf
pydantic
python-multipart
httpx
xgboost
scikit-learn
joblib
chromadb
sentence-transformers
langchain
pypdf
langchain-community
google-generativeai
python-dotenv
pandas
numpy
